apps/worker/requirements.txt
supabase==2.3.0
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
openai==1.3.0
apscheduler==3.10.4
ffmpeg-python==0.2.0
google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
pydantic==2.5.0
pillow==10.1.0
numpy==1.26.0

